Refactor: Move _Annotation from _stbt.core to _stbt.logging
This removes a dependency from stbt.match on _stbt.core.
